+++ src/main/java/lotto/Application.java
@@ -1,9 +1,10 @@
 package lotto;
 
-import lotto.model.Lotto;
+import lotto.model.*;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Application {
@@ -11,29 +12,46 @@ public class Application {
     public static void main(String[] args) {
 
         InputView inputView = new InputView();
+        LottoGame lottoGame = new LottoGame();
         ResultView resultView = new ResultView();
-        HitLotto hitLotto = new HitLotto();
 
-        // 구입금액 입력
         long purchaseAmount = inputView.purchaseAmount();
+        int manualPurchaseCount = inputView.manualPurchaseCount();
+        int autoPurchaseCount = lottoGame.getAutoPurchaseCount(purchaseAmount, manualPurchaseCount);
+        String[] manualLottos = inputView.inputManualLotto(manualPurchaseCount);
 
-        // 로또 생성
-        Lotto lotto = new Lotto(purchaseAmount, hitLotto.getLotto(hitLotto.getPurchaseCount(purchaseAmount)));
 
-        // 구매내역 출력
-        resultView.purchaseHistory(hitLotto.getPurchaseCount(purchaseAmount), lotto);
+        // 수동 번호 생성
+        List<Lotto> lottos = new ArrayList<>();
+        for (int i = 0; i < manualLottos.length; i++) {
+            Lotto lotto = lottoGame.createManualLotto(manualLottos[i]);
+            lottos.add(lotto);
+        }
+
+        resultView.purchaseLottoCount(manualPurchaseCount, autoPurchaseCount);
+
+        // 자동 번호 생성
+        if (autoPurchaseCount > 0) {
+            for (int i = 0; i < autoPurchaseCount; i++) {
+                Lotto lotto = lottoGame.createAutoLotto();
+                resultView.printToAutoLotto(lotto);
+                lottos.add(lotto);
+            }
+        }
+
+        UserLotto userLotto = new UserLotto(lottos);
+
 
         // 지난 주 당첨 번호 입력
         String stringWinningNumbers = inputView.inputWinningNumber();
-        List<Integer> winningNumbers = hitLotto.convertWinningNumbers(stringWinningNumbers);
+        Lotto winningNumber = lottoGame.createManualLotto(stringWinningNumbers);
 
         // 보너스 볼 입력
         int bonusNumber = inputView.inputBonusNumber();
 
-        // 로또 당첨 확인
-        lotto.confirmWinningNumber(winningNumbers, bonusNumber);
+        WinningLotto winningLotto = new WinningLotto(winningNumber, bonusNumber);
 
-        // 당첨 통계 출력
-        resultView.printToWinningStatistics(lotto);
+        // 당첨통계
+        resultView.printToWinningStatistics2(userLotto, winningLotto, purchaseAmount);
     }
 }
