+++ src/main/java/lotto/model/Level.java
@@ -3,14 +3,12 @@ package lotto.model;
 import java.util.Arrays;
 
 public enum Level {
-    FIRST(6, 2000000000),
-    SECOND(5, 1500000),
-    THIRD(4, 50000),
-    FOURTH(3, 5000),
-    FIFTH(2, 0),
-    SIXTH(1, 0),
-    SEVENTH(0, 0),
-    BONUS(7, 30000000);
+    FIRST(1, 2000000000),
+    SECOND(2, 30000000),
+    THIRD(3, 1500000),
+    FOURTH(4, 50000),
+    FIFTH(5, 5000),
+    SIXTH(0, 0);
 
     private final long value;
     private final long amount;
