+++ src/main/java/lotto/model/LottoNumber.java
@@ -1,33 +1,45 @@
 package lotto.model;
 
-import java.util.List;
+import java.util.Objects;
 
 public class LottoNumber {
 
-    private List<Integer> numbers;
+    private static final int MINIMUM_NUMBER = 1;
+    private static final int MAXIMUM_NUMBER = 45;
 
-    private Level level;
+    private int value;
 
-    public LottoNumber(List<Integer> numbers, Level level) {
-        this.numbers = numbers;
-        this.level = level;
+    public LottoNumber(String value) {
+
+        this(Integer.parseInt(value));
     }
 
-    public List<Integer> getNumbers() {
-        return numbers;
+    public LottoNumber(int value) {
+        if (value < MINIMUM_NUMBER || value > MAXIMUM_NUMBER) {
+            throw new IllegalArgumentException("1 ~ 45 사이의 숫자만 가능");
+        }
+        this.value = value;
     }
 
-    public void setLevel(Level level) {
-        this.level = level;
+    public int getValue() {
+        return value;
     }
 
-    public Level getLevel() { return level; }
-
-    public String printNumber() {
-
-        return numbers.toString();
+    @Override
+    public String toString() {
+        return Integer.toString(value);
     }
-    public long getWinningAmount() {
-        return level.getAmount();
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        LottoNumber that = (LottoNumber) o;
+        return value == that.value;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
     }
 }
