--- src/main/java/lotto/model/LottoNumber2.java
@@ -1,45 +1,0 @@
-package lotto.model;
-
-import java.util.Objects;
-
-public class LottoNumber2 {
-
-    private static final int MINIMUM_NUMBER = 1;
-    private static final int MAXIMUM_NUMBER = 45;
-
-    private int value;
-
-    public LottoNumber2(String value) {
-
-        this(Integer.parseInt(value));
-    }
-
-    public LottoNumber2(int value) {
-        if (value < MINIMUM_NUMBER || value > MAXIMUM_NUMBER) {
-            throw new IllegalArgumentException("1 ~ 45 사이의 숫자만 가능");
-        }
-        this.value = value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    @Override
-    public String toString() {
-        return Integer.toString(value);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        LottoNumber2 that = (LottoNumber2) o;
-        return value == that.value;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-}
