+++ src/main/java/lotto/model/UserLotto.java
@@ -5,20 +5,20 @@ import java.util.List;
 
 public class UserLotto {
 
-    private List<Lotto2> lottos;
+    private List<Lotto> lottos;
 
-    public UserLotto(List<Lotto2> lottos) {
+    public UserLotto(List<Lotto> lottos) {
         this.lottos = lottos;
     }
 
-    public int matchCount(WinningLotto winningLotto, Level2 level) {
+    public int matchCount(WinningLotto winningLotto, Level level) {
         return (int) lottos.stream()
                 .filter(it -> it.match(winningLotto) == level.getValue())
                 .count();
     }
 
     public double getTotAteOfReturn(WinningLotto winningLotto, long purchaseAmount) {
-        long totalWinningAmount = Arrays.stream(Level2.values())
+        long totalWinningAmount = Arrays.stream(Level.values())
                 .map(it -> matchCount(winningLotto, it) * it.getAmount())
                 .reduce(0L, Long::sum);
