+++ src/main/java/lotto/model/WinningLotto.java
@@ -2,16 +2,16 @@ package lotto.model;
 
 public class WinningLotto {
 
-    private Lotto2 lotto;
+    private Lotto lotto;
 
     private int bonusNumber;
 
-    public WinningLotto(Lotto2 lotto, int bonusNumber) {
+    public WinningLotto(Lotto lotto, int bonusNumber) {
         this.lotto = lotto;
         this.bonusNumber = bonusNumber;
     }
 
-    public Lotto2 getLotto() {
+    public Lotto getLotto() {
         return lotto;
     }
