+++ src/test/java/lotto/LottoTest.java
@@ -1,7 +1,6 @@
 package lotto;
 
-import lotto.model.Lotto2;
-import lotto.model.LottoNumber2;
+import lotto.model.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -49,13 +48,13 @@ public class LottoTest {
     @DisplayName(value = "로또 번호 생성")
     @Test
     void createLottoNumbers2() {
-        List<LottoNumber2> numbers = IntStream.rangeClosed(MINIMUM_NUMBER, MAXIMUM_NUMBER)
+        List<LottoNumber> numbers = IntStream.rangeClosed(MINIMUM_NUMBER, MAXIMUM_NUMBER)
                 .boxed()
-                .map(it -> new LottoNumber2(it))
+                .map(it -> new LottoNumber(it))
                 .collect(Collectors.toList());
 
         Collections.shuffle(numbers);
-        List<LottoNumber2> lotto = numbers.subList(0, 6);
+        List<LottoNumber> lotto = numbers.subList(0, 6);
         Collections.sort(lotto, (a, b) -> a.getValue() - b.getValue());
 
         assertThat(lotto.size()).isSameAs(6);
@@ -66,9 +65,9 @@ public class LottoTest {
     void createManualLottoNumber2() {
         String manualLotto = "1, 3, 5, 6, 7, 8";
 
-        List<LottoNumber2> numbers = Arrays.stream(manualLotto.split(","))
+        List<LottoNumber> numbers = Arrays.stream(manualLotto.split(","))
                 .map(String::trim)
-                .map(it -> new LottoNumber2(it))
+                .map(it -> new LottoNumber(it))
                 .collect(Collectors.toList());
 
         assertThat(numbers.size()).isSameAs(6);
@@ -90,29 +89,5 @@ public class LottoTest {
                 .filter(it -> winningNumbers.contains(it))
                 .count();
         assertThat(count2).isSameAs(3L);
-    }
-
-    @Test
-    void temp() {
-        String input = "1, 2, 3, 4, 5, 6";
-        int[] values = Arrays.stream(input.split(", "))
-                .mapToInt(Integer::parseInt)
-                .toArray();
-
-        for (int value : values) {
-            System.out.println(value);
-        }
-
-        List<Integer> results = Arrays.stream(input.split(", "))
-                .mapToInt(Integer::parseInt)
-                .boxed()
-                .collect(Collectors.toList());
-
-//        List<Integer> results = Arrays.stream(values).boxed()
-//                .collect(Collectors.toList());
-
-        for (Integer result : results) {
-            System.out.println(result);
-        }
     }
 }
